cli: move show_patch() to diff_util
I think this belongs to the same category as show_diff().
